Eliminate browser browser module
