+++ js/util/browser.js
@@ -1,29 +1,46 @@
 'use strict';
 
-var window = require('./window');
-
-exports.window = window;
-exports.document = window.document;
-/*
- * When browserify builds Mapbox GL JS, it redirects all require() statements
- * from this file, js/util/browser.js, to js/util/browser/browser.js.
- * The latter relies on running in a real browser: 'window' must be defined,
- * as well as other browser-specific globals. This file, on the other hand,
- * is comfortable running under node.js, which is why it's the default require:
- * it's used for tests.
+/**
+ * Unlike js/util/browser.js, this code is written with the expectation
+ * of a browser environment with a global 'window' object
+ * @module browser
+ * @private
  */
 
+var window = require('./window');
+
+/**
+ * Provides a function that outputs milliseconds: either performance.now()
+ * or a fallback to Date.now()
+ */
+module.exports.now = (function() {
+    if (window.performance &&
+        window.performance.now) {
+        return window.performance.now.bind(window.performance);
+    } else {
+        return Date.now.bind(Date);
+    }
+}());
+
+var frame = window.requestAnimationFrame ||
+    window.mozRequestAnimationFrame ||
+    window.webkitRequestAnimationFrame ||
+    window.msRequestAnimationFrame;
+
 exports.frame = function(fn) {
-    return setImmediate(fn);
+    return frame(fn);
 };
+
+var cancel = window.cancelAnimationFrame ||
+    window.mozCancelAnimationFrame ||
+    window.webkitCancelAnimationFrame ||
+    window.msCancelAnimationFrame;
 
 exports.cancelFrame = function(id) {
-    return clearImmediate(id);
+    cancel(id);
 };
 
-module.exports.now = Date.now.bind(Date);
-
-exports.timed = function(fn, dur, ctx) {
+exports.timed = function (fn, dur, ctx) {
     if (!dur) {
         fn.call(ctx, 1);
         return null;
@@ -49,11 +66,28 @@ exports.timed = function(fn, dur, ctx) {
     return function() { abort = true; };
 };
 
-exports.supported = function () {
-    return true;
-};
+/**
+ * Test if the current browser supports Mapbox GL JS
+ * @param {Object} options
+ * @param {boolean} [options.failIfMajorPerformanceCaveat=false] Return `false`
+ *   if the performance of Mapbox GL JS would be dramatically worse than
+ *   expected (i.e. a software renderer would be used)
+ * @return {boolean}
+ */
+exports.supported = require('mapbox-gl-supported');
 
-exports.devicePixelRatio = 1;
-exports.hardwareConcurrency = 8;
+exports.hardwareConcurrency = window.navigator.hardwareConcurrency || 4;
+
+Object.defineProperty(exports, 'devicePixelRatio', {
+    get: function() { return window.devicePixelRatio; }
+});
+
 exports.supportsWebp = false;
-exports.supportsGeolocation = false;
+
+var webpImgTest = window.document.createElement('img');
+webpImgTest.onload = function() {
+    exports.supportsWebp = true;
+};
+webpImgTest.src = 'data:image/webp;base64,UklGRh4AAABXRUJQVlA4TBEAAAAvAQAAAAfQ//73v/+BiOh/AAA=';
+
+exports.supportsGeolocation = !!window.navigator.geolocation;
