+++ js/util/window.js
@@ -1,4 +1,12 @@
 'use strict';
 
 var jsdom = require('jsdom');
-module.exports = jsdom.jsdom().defaultView;
+
+var window = jsdom.jsdom().defaultView;
+
+window.requestAnimationFrame = function(callback) { return setTimeout(callback, 0); };
+window.cancelAnimationFrame = clearTimeout;
+
+window.devicePixelRatio = 1;
+
+module.exports = window;
