+++ package.json
@@ -76,12 +76,12 @@
     "vinyl": "1.1.1",
     "vinyl-fs": "2.4.3",
     "watchify": "^3.7.0",
+    "webgl-mock": "^0.1.6",
     "webpack": "^1.13.1",
     "webworkify-webpack": "^1.1.3"
   },
   "browser": {
     "./js/util/ajax.js": "./js/util/browser/ajax.js",
-    "./js/util/browser.js": "./js/util/browser/browser.js",
     "./js/util/window.js": "./js/util/browser/window.js",
     "./js/util/canvas.js": "./js/util/browser/canvas.js",
     "./js/util/web_worker.js": "./js/util/browser/web_worker.js"
